replace Double/Float wrapper types with primitives in node payoffs and pot fields

+++ riversolver/src/main/java/icybee/solver/trainable/CfrPlusTrainable.java
@@ -111,7 +111,7 @@ public class CfrPlusTrainable extends Trainable {
     public float[] getcurrentStrategy(int private_id) {
         float[] retval = new float[this.action_number];
         if (this.r_plus_sum == null || this.r_plus_sum[private_id] == 0) {
-            Arrays.fill(retval, Float.valueOf(1) / this.action_number);
+            Arrays.fill(retval, 1F / this.action_number);
         } else {
             for (int action_id = 0; action_id < action_number; action_id++) {
                 int index = action_id * this.card_number + private_id;
